% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-utils.R
\name{copy_file}
\alias{copy_file}
\title{Copy directories recursively, creating a new directory if not already there}
\usage{
copy_file(from, to, overwrite = TRUE)
}
\arguments{
\item{string}{string}
}
\value{
string
}
\description{
copy_file
}
\examples{
\dontrun{
}
}

