% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-utils.R
\name{file_path_sans_ext}
\alias{file_path_sans_ext}
\title{file_path_sans_ext}
\usage{
file_path_sans_ext(x)
}
\arguments{
\item{string}{string}
}
\value{
string
}
\description{
borrowed from tools package
}
\examples{
\dontrun{
}
}

