% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_multilines}
\alias{hgch_multilines}
\title{hgch_multilines}
\usage{
hgch_multilines(fchrono, type = "column", title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL)
}
\description{
hgch_multilines
}
\section{ftype}{
 Ye-Nu*
}

