% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_multilines_ynp}
\alias{hgch_multilines_ynp}
\title{hgch_multilines_ynp}
\usage{
hgch_multilines_ynp(f, type = "column", title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL)
}
\description{
hgch_multilines_ynp
}
\section{ftype}{
 Ye-Nu*
}

