% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-utils.R
\name{list_dirs}
\alias{list_dirs}
\title{List directories in a directory}
\usage{
list_dirs(path = ".", recursive = TRUE, fullPath = TRUE)
}
\arguments{
\item{string}{string}
}
\value{
string
}
\description{
list_dirs
}
\examples{
\dontrun{
}
}

