% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-utils.R
\name{read_sysfile}
\alias{read_csv_sysfile}
\alias{read_sysfile}
\title{Read contents of a system file into a character string}
\usage{
read_sysfile(..., package = "mwr")

read_csv_sysfile(path)
}
\arguments{
\item{string}{string}

\item{string}{string}
}
\value{
string

string
}
\description{
read_sysfile

read_sysfile
}
\examples{
\dontrun{
}
\dontrun{
}
}

